@@ -1,70 +1,44 @@
 name: contactos
-description: A Flutter plugin to retrieve and manage contacts on Android and iOS devices.
-version: 1.0.0
-
-homepage: https://github.com/ziqq/contactos
-
-repository: https://github.com/ziqq/contactos
-
-issue_tracker: https://github.com/ziqq/contactos/issues
+description: Android and iOS implementation of the contactos plugin.
+repository: https://github.com/ziqq/contactos/contactos
+issue_tracker: https://github.com/ziqq/contactos/issues?q=is%3Aissue+is%3Aopen+label%3A%22p%3A+contactos%22
+version: 1.0.0-pre.1
 
 publish_to: 'none'
 
-funding:
-  - https://www.buymeacoffee.com/ziqq
-  - https://www.patreon.com/ziqq
-  - https://boosty.to/ziqq
-
-platforms:
-  android:
-  ios:
-
-#screenshots:
-#  - description: 'Example of using the library to check app version.'
-#    path: example.png
+environment:
+  sdk: '>=3.6.0 <4.0.0'
+  flutter: ">=3.29.3"
 
 flutter:
   plugin:
+    implements: contactos
     platforms:
       android:
-        default_package: contactos_android
+        package: flutter.plugins.contactos
+        pluginClass: ContactosPlugin
+        dartPluginClass: Contactos
       ios:
-        default_package: contactos_ios
-
-environment:
-  sdk: '>=3.6.0 <4.0.0'
-  flutter: ">=3.27.1"
+        pluginClass: ContactosPlugin
+        dartPluginClass: Contactos
+      macos:
+        pluginClass: ContactosPlugin
+        dartPluginClass: Contactos
 
 dependencies:
   flutter:
     sdk: flutter
-
-  # Utilities
-  collection: '>=1.19.0 <2.0.0'
-  meta: '>=1.15.0 <2.0.0'
-
-  # contactos_android: ^1.0.0
-  contactos_android:
-    path: ../contactos_android
-
-  # contactos_ios: ^1.0.0
-  contactos_ios:
-    path: ../contactos_ios
 
   # contactos_platform_interface: ^1.0.0
   contactos_platform_interface:
     path: ../contactos_platform_interface
 
 dev_dependencies:
-  # Integration tests for Flutter
-  integration_test:
-    sdk: flutter
-  # Unit & Widget tests for Flutter
   flutter_test:
     sdk: flutter
 
-  # Logging
-  l: ^5.0.0
-
   # Linting
   flutter_lints: ^5.0.0
+
+topics:
+  - contacts
