@@ -4,7 +4,11 @@ import 'package:contactos/contactos.dart';
 import 'package:contactos_example/main.dart';
 import 'package:flutter/material.dart';
 
+/// {@template navite_contacts_picker_screen}
+/// NativeContactsPickerScreen widget.
+/// {@endtemplate}
 class NativeContactsPickerScreen extends StatefulWidget {
+  /// {@macro navite_contacts_picker_screen}
   const NativeContactsPickerScreen({super.key});
 
   @override
@@ -23,7 +27,7 @@ class _NativeContactsPickerScreenState
 
   Future<void> _pickContact() async {
     try {
-      final Contact? contact = await Contactos.openDeviceContactPicker(
+      final contact = await Contactos.instance.openDeviceContactPicker(
         iOSLocalizedLabels: iOSLocalizedLabels,
       );
       setState(() => _contact = contact);
