@@ -1,8 +1,9 @@
 // This is a basic Flutter widget test.
-// To perform an interaction with a widget in your test, use the WidgetTester utility that Flutter
-// provides. For example, you can send tap and scroll gestures. You can also use WidgetTester to
-// find child widgets in the widget tree, read text, and verify that the values of widget properties
-// are correct.
+// To perform an interaction with a widget in your test,
+// use the WidgetTester utility that Flutter
+// provides. For example, you can send tap and scroll gestures.
+// You can also use WidgetTester to find child widgets in the widget tree,
+// read text, and verify that the values of widget properties are correct.
 
 import 'package:contactos_example/src/screens/contacts_list_screen.dart';
 import 'package:flutter/material.dart';
