@@ -1,213 +1,139 @@
-// ignore_for_file: sort_constructors_first
+// Copyright 2025 Anton Ustinoff<a.a.ustinoff@gmail.com>. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
 
-import 'dart:async';
+import 'package:contactos_platform_interface/contactos_platform_interface.dart';
+import 'package:flutter/foundation.dart';
 
-import 'package:contactos/contactos_platform_interface/lib/exception.dart';
-import 'package:contactos/contactos_platform_interface/lib/model.dart';
-import 'package:contactos/contactos_platform_interface/lib/typedef.dart';
-import 'package:flutter/services.dart';
+/// The iOS implementation of [ContactosPlatform].
+///
+/// This class implements the `package:contactos` functionality for iOS.
+class Contactos extends ContactosPlatform {
+  Contactos._({@visibleForTesting MethodChannelContactos? channel})
+      : channel = channel ?? MethodChannelContactos.instance,
+        super();
 
-export 'share.dart';
+  /// Underlying channel-based implementation.
+  final MethodChannelContactos channel;
 
-/// {@template contactos}
-/// A service that provides access to the device contacts.
-/// {@endtemplate}
-final class Contactos {
-  /// {@macro contactos}
-  const Contactos._();
+  static Contactos? _instance;
 
-  // static final Contactos instance = Contactos._internal();
-  // factory Contactos() => instance;
+  /// Returns the [Contactos] singleton instance.
+  /// Also registers this as the default platform implementation.
+  // ignore: prefer_constructors_over_static_methods
+  static Contactos get instance => _instance ??= Contactos._();
 
-  /// {@macro contactos}
-  // Contactos._internal();
+  /// Registers this class as the default instance of [Contactos].
+  static void registerWith() {
+    ContactosPlatform.instance = Contactos.instance;
+  }
 
-  /// The [MethodChannel] used to interact with the platform side of the plugin.
-  static const MethodChannel _channel =
-      MethodChannel('github.com/ziqq/contactos');
+  @protected
+  @override
+  ContactosPlatform delegateFor({
+    required MethodChannelContactos channel,
+  }) =>
+      Contactos._(channel: channel);
 
-  /// Fetches all contacts, or when specified, the contacts with a name
-  /// matching [query]
-  static Future<List<Contact>> getContacts({
+  @override
+  Future<void> addContact(Contact c) => channel.addContact(c);
+
+  @override
+  Future<void> deleteContact(Contact c) => channel.deleteContact(c);
+
+  @override
+  Future<void> updateContact(Contact c) => channel.updateContact(c);
+
+  @override
+  Future<Uint8List?> getAvatar(
+    Contact contact, {
+    bool photoHighRes = true,
+  }) =>
+      channel.getAvatar(contact, photoHighRes: photoHighRes);
+
+  @override
+  Future<List<Contact>> getContacts({
     String? query,
     bool withThumbnails = true,
     bool photoHighResolution = true,
     bool orderByGivenName = true,
     bool iOSLocalizedLabels = true,
     bool androidLocalizedLabels = true,
-  }) async {
-    final contacts = await _channel.invokeMethod(
-      'getContacts',
-      <String, dynamic>{
-        'query': query,
-        'withThumbnails': withThumbnails,
-        'photoHighResolution': photoHighResolution,
-        'orderByGivenName': orderByGivenName,
-        'iOSLocalizedLabels': iOSLocalizedLabels,
-        'androidLocalizedLabels': androidLocalizedLabels,
-      },
-    );
-    return (contacts as Iterable<dynamic>)
-        .map((c) => Contact.fromMap(c as JSON))
-        .toList();
-  }
+  }) =>
+      channel.getContacts(
+        query: query,
+        withThumbnails: withThumbnails,
+        photoHighResolution: photoHighResolution,
+        orderByGivenName: orderByGivenName,
+        iOSLocalizedLabels: iOSLocalizedLabels,
+        androidLocalizedLabels: androidLocalizedLabels,
+      );
 
-  /// Fetches all contacts, or when specified, the contacts with the phone
-  /// matching [phone]
-  static Future<List<Contact>> getContactsForPhone(
+  @override
+  Future<List<Contact>> getContactsForEmail(
+    String email, {
+    bool withThumbnails = true,
+    bool photoHighResolution = true,
+    bool orderByGivenName = true,
+    bool iOSLocalizedLabels = true,
+    bool androidLocalizedLabels = true,
+  }) =>
+      channel.getContactsForEmail(
+        email,
+        withThumbnails: withThumbnails,
+        photoHighResolution: photoHighResolution,
+        orderByGivenName: orderByGivenName,
+        iOSLocalizedLabels: iOSLocalizedLabels,
+        androidLocalizedLabels: androidLocalizedLabels,
+      );
+
+  @override
+  Future<List<Contact>> getContactsForPhone(
     String? phone, {
     bool withThumbnails = true,
     bool photoHighResolution = true,
     bool orderByGivenName = true,
     bool iOSLocalizedLabels = true,
     bool androidLocalizedLabels = true,
-  }) async {
-    if (phone == null || phone.isEmpty) return List.empty();
-
-    final contacts =
-        await _channel.invokeMethod('getContactsForPhone', <String, dynamic>{
-      'phone': phone,
-      'withThumbnails': withThumbnails,
-      'photoHighResolution': photoHighResolution,
-      'orderByGivenName': orderByGivenName,
-      'iOSLocalizedLabels': iOSLocalizedLabels,
-      'androidLocalizedLabels': androidLocalizedLabels,
-    });
-    return (contacts as Iterable<dynamic>)
-        .map((c) => Contact.fromMap(c as JSON))
-        .toList();
-  }
-
-  /// Fetches all contacts, or when specified, the contacts with the email
-  /// matching [email]
-  /// Works only on iOS
-  static Future<List<Contact>> getContactsForEmail(String email,
-      {bool withThumbnails = true,
-      bool photoHighResolution = true,
-      bool orderByGivenName = true,
-      bool iOSLocalizedLabels = true,
-      bool androidLocalizedLabels = true}) async {
-    final contacts = await _channel.invokeMethod(
-      'getContactsForEmail',
-      <String, dynamic>{
-        'email': email,
-        'withThumbnails': withThumbnails,
-        'photoHighResolution': photoHighResolution,
-        'orderByGivenName': orderByGivenName,
-        'iOSLocalizedLabels': iOSLocalizedLabels,
-        'androidLocalizedLabels': androidLocalizedLabels,
-      },
-    );
-    if (contacts is! List) return const [];
-    return contacts.map((c) => Contact.fromMap(c as JSON)).toList();
-  }
-
-  /// Loads the avatar for the given contact and returns it. If the user does
-  /// not have an avatar, then `null` is returned in that slot. Only implemented
-  /// on Android.
-  static Future<Uint8List?> getAvatar(
-    final Contact contact, {
-    final bool photoHighRes = true,
   }) =>
-      _channel.invokeMethod(
-        'getAvatar',
-        <String, dynamic>{
-          'contact': contact.toMap(),
-          'identifier': contact.identifier,
-          'photoHighResolution': photoHighRes,
-        },
+      channel.getContactsForPhone(
+        phone,
+        withThumbnails: withThumbnails,
+        photoHighResolution: photoHighResolution,
+        orderByGivenName: orderByGivenName,
+        iOSLocalizedLabels: iOSLocalizedLabels,
+        androidLocalizedLabels: androidLocalizedLabels,
       );
 
-  /// Adds the [contact] to the device contact list
-  static Future<void> addContact(Contact contact) =>
-      _channel.invokeMethod('addContact', contact.toMap());
-
-  /// Deletes the [contact] if it has a valid identifier
-  static Future<void> deleteContact(Contact contact) =>
-      _channel.invokeMethod('deleteContact', contact.toMap());
-
-  /// Updates the [contact] if it has a valid identifier
-  static Future<void> updateContact(Contact contact) =>
-      _channel.invokeMethod('updateContact', contact.toMap());
-
-  /// Opens the contact form with the fields prefilled with the values from the
-  static Future<Contact> openContactForm({
+  @override
+  Future<Contact> openContactForm({
     bool iOSLocalizedLabels = true,
     bool androidLocalizedLabels = true,
-  }) async {
-    final result = await _channel.invokeMethod(
-      'openContactForm',
-      <String, dynamic>{
-        'iOSLocalizedLabels': iOSLocalizedLabels,
-        'androidLocalizedLabels': androidLocalizedLabels,
-      },
-    );
-    return _handleFormOperation(result);
-  }
+  }) =>
+      channel.openContactForm(
+        iOSLocalizedLabels: iOSLocalizedLabels,
+        androidLocalizedLabels: androidLocalizedLabels,
+      );
 
-  /// Opens the contact form with the fields prefilled with the values from the
-  /// [contact] parameter
-  static Future<Contact> openExistingContact(
+  @override
+  Future<Contact?> openDeviceContactPicker({
+    bool iOSLocalizedLabels = true,
+    bool androidLocalizedLabels = true,
+  }) =>
+      channel.openDeviceContactPicker(
+        iOSLocalizedLabels: iOSLocalizedLabels,
+        androidLocalizedLabels: androidLocalizedLabels,
+      );
+
+  @override
+  Future<Contact> openExistingContact(
     Contact contact, {
     bool iOSLocalizedLabels = true,
     bool androidLocalizedLabels = true,
-  }) async {
-    dynamic result = await _channel.invokeMethod(
-      'openExistingContact',
-      <String, dynamic>{
-        'contact': contact.toMap(),
-        'iOSLocalizedLabels': iOSLocalizedLabels,
-        'androidLocalizedLabels': androidLocalizedLabels,
-      },
-    );
-    return _handleFormOperation(result);
-  }
-
-  // Displays the device/native contact picker dialog and returns the contact selected by the user
-  static Future<Contact?> openDeviceContactPicker({
-    bool iOSLocalizedLabels = true,
-    bool androidLocalizedLabels = true,
-  }) async {
-    var result = await _channel.invokeMethod(
-      'openDeviceContactPicker',
-      <String, dynamic>{
-        'iOSLocalizedLabels': iOSLocalizedLabels,
-        'androidLocalizedLabels': androidLocalizedLabels,
-      },
-    );
-    // result contains either :
-    // - an List of contacts containing 0 or 1 contact
-    // - a FormOperationErrorCode value
-    if (result is List) {
-      if (result.isEmpty) return null;
-      result = result.first;
-    }
-    return _handleFormOperation(result);
-  }
-
-  // ignore: avoid_annotating_with_dynamic
-  static Contact _handleFormOperation(dynamic result) {
-    if (result is int) {
-      switch (result) {
-        case 1:
-          throw const FormOperationException(
-            errorCode: FormOperationErrorCode.FORM_OPERATION_CANCELED,
-          );
-        case 2:
-          throw const FormOperationException(
-            errorCode: FormOperationErrorCode.FORM_COULD_NOT_BE_OPEN,
-          );
-        default:
-          throw const FormOperationException(
-            errorCode: FormOperationErrorCode.FORM_OPERATION_UNKNOWN_ERROR,
-          );
-      }
-    } else if (result is JSON) {
-      return Contact.fromMap(result);
-    } else {
-      throw const FormOperationException(
-        errorCode: FormOperationErrorCode.FORM_OPERATION_UNKNOWN_ERROR,
+  }) =>
+      channel.openExistingContact(
+        contact,
+        iOSLocalizedLabels: iOSLocalizedLabels,
+        androidLocalizedLabels: androidLocalizedLabels,
       );
-    }
-  }
 }
